# src/api/content-api-openapi.yaml
openapi: 3.0.3
info:
  title: MLG.clan Content Storage and Retrieval API
  description: |
    Comprehensive gaming platform API for content storage, retrieval, and management.
    
    ## Features
    - Content CRUD operations for gaming clips, screenshots, guides, and reviews
    - MLG token integration with Solana blockchain voting system
    - Advanced search and filtering capabilities
    - Content moderation and reporting system
    - Real-time analytics and metrics tracking
    - Scalable file upload/download handling
    - User content management and statistics
    
    ## Authentication
    - JWT tokens for user authentication
    - Solana wallet signature verification for blockchain operations
    - Rate limiting and abuse prevention
    
    ## MLG Token Integration
    - Contract Address: 7XJiwLDrjzxDYdZipnJXzpr1iDTmK55XixSFAa7JgNEL
    - Vote costs: Upvote (1 MLG), Downvote (2 MLG), Super Vote (5 MLG)
    - Reward multipliers based on content type and engagement
    
  version: 1.0.0
  contact:
    name: MLG.clan Development Team
    url: https://mlg.clan/api-support
    email: api-support@mlg.clan
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.mlg.clan/v1
    description: Production API Server
  - url: https://staging-api.mlg.clan/v1
    description: Staging API Server
  - url: http://localhost:3000/api/v1
    description: Local Development Server
security:
  - BearerAuth: []
  - WalletSignature: []
paths:
...
  /content/{contentId}/votes:
    get:
      summary: Get content vote history
      description: |
        Retrieve vote history and statistics for content.
        Includes vote breakdown by type and top voters.
      tags:
        - Voting
      security:
        - BearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/ContentId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: voteType
          in: query
          description: Filter by vote type
          schema:
            $ref: '#/components/schemas/VoteType'
        - name: includeAnonymous
          in: query
          description: Include anonymous votes
          schema:
            type: boolean
            default: true
        - name: sortBy
          in: query
          description: Sort votes by field
          schema:
            type: string
            enum: [created_at, token_amount]
            default: created_at
      responses:
        '200':
          description: Vote history retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          votes:
                            type: array
                            items:
                              $ref: '#/components/schemas/Vote'
                          summary:
                            type: object
                            properties:
                              totalVotes:
                                type: integer
                              totalTokensBurned:
                                type: number
                                format: float
                              voteBreakdown:
                                type: object
                                properties:
                                  upvotes:
                                    type: integer
                                  downvotes:
                                    type: integer
                                  superVotes:
                                    type: integer
                              topVoters:
                                type: array
                                items:
                                  type: object
                                  properties:
                                    userId:
                                      type: string
                                    username:
                                      type: string
                                    totalTokensBurned:
                                      type: number
                                      format: float
                                    voteCount:
                                      type: integer
                          pagination:
                            $ref: '#/components/schemas/PaginationResponse'

  # Search and Discovery
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for user authentication
    
    WalletSignature:
      type: apiKey
      in: header
      name: X-Wallet-Signature
      description: |
        Solana wallet signature for blockchain operations.
        Required format: {signature}:{message}:{publicKey}
  parameters:
    ContentId:
      name: contentId
      in: path
      required: true
      description: Content UUID
      schema:
        type: string
        format: uuid
        example: "123e4567-e89b-12d3-a456-426614174000"
    Limit:
      name: limit
      in: query
      description: Number of items to return
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      description: Number of items to skip
      schema:
        type: integer
        minimum: 0
        default: 0
...
  schemas:
...
    Vote:
      type: object
      required:
        - id
        - contentId
        - userId
        - walletAddress
        - voteType
        - tokensBurned
        - transactionSignature
        - walletSignature
        - createdAt
        - blockTime
      properties:
        id:
          type: string
          format: uuid
        contentId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        walletAddress:
          type: string
          pattern: '^[1-9A-HJ-NP-Za-km-z]{32,44}$'
        voteType:
          $ref: '#/components/schemas/VoteType'
        tokensBurned:
          type: number
          format: float
          minimum: 0.1
          example: 1.0
        transactionSignature:
          type: string
          pattern: '^[1-9A-HJ-NP-Za-km-z]{87,88}$'
          example: "5J8QvU7snqjBxNqVQhGjPFzQFzQYU7snqjBxNqVQhGjPFzQFzQYU7snqjBxNqVQhGjPFzQFzQY"
        walletSignature:
          $ref: '#/components/schemas/WalletSignature'
        comment:
          type: string
          maxLength: 200
        isAnonymous:
          type: boolean
          default: false
        createdAt:
          type: string
          format: date-time
        blockTime:
          type: string
          format: date-time
          description: Solana blockchain timestamp
    WalletSignature:
      type: object
      required:
        - signature
        - message
        - publicKey
        - verified
      properties:
        signature:
          type: string
          example: "signature-data-here"
        message:
          type: string
          example: "MLG.clan Vote Verification"
        publicKey:
          type: string
          pattern: '^[1-9A-HJ-NP-Za-km-z]{32,44}$'
        verified:
          type: boolean
          example: true
...
    VoteType:
      type: string
      enum:
        - upvote
        - downvote
        - super_vote
      example: upvote
...
    SuccessResponse:
      type: object
      required:
        - success
        - data
        - meta
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          description: Response data
        meta:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
            requestId:
              type: string
              format: uuid
            version:
              type: string
              example: "1.0.0"
...
    PaginationResponse:
      type: object
      required:
        - total
        - limit
        - offset
        - hasNext
        - hasPrev
      properties:
        total:
          type: integer
          minimum: 0
          example: 150
        limit:
          type: integer
          minimum: 1
          example: 20
        offset:
          type: integer
          minimum: 0
          example: 40
        hasNext:
          type: boolean
          example: true
        hasPrev:
          type: boolean
          example: true
        nextCursor:
          type: string
          example: "eyJpZCI6IjEyMyIsInRzIjoxNjQwOTk1MjAwfQ"
        prevCursor:
          type: string
          example: "eyJpZCI6IjEwMCIsInRzIjoxNjQwOTk1MTAwfQ"
...